Model of hybrid classical-quantum computing method
Here, a model of classical computing that uses an electric field is proposed. The model can be used to create a wireless logic circuit. In this model, an electric field is used to transfer bit signals. The transfer medium can be a semiconductor, an optical fiber, some other material, or a combination.