The global economic outlook
 ~ for the steady economic growth of the world
This paper explores the need to establish a world federal government as the next world governing framework from the broadest economic perspective, by focusing on the relationship between private sector production rate and two factors, economic growth rate and economic stability. A redistribution method is considered as financial policy, and the issuing of world legal tender is considered as monetary policy, which attempts to include a redistribution system covered by a transfer fee.
This paper is an outlook of global economics and provides fundamental solutions to the current economic problems with an upcoming technology, blockchains.