Nano two-state constant modules using graphene sheets
In the expected future, if we can create plane single-layer graphene sheets and can connect them vertically or in parallel by some means, the design presented in the following figures would make it possible to produce nano-scale physical properties, opened and closed state. As the result, they could be parts of nano-scale filters, electronic devices, or any nanoscale fine-tunings parameters.