Nano two-state constant modules using graphene sheets
In future, if we can create plane single-layer graphene sheets and connect them vertically or in parallel by some means, the designs presented in the following figures would make it possible to produce objects with nano-scale physical properties, in both opened and closed states. As a result, they could be parts of nano-scale filters, electronic devices, or any nanoscale fine-tuning parameters.