The formulation of prime numbers
Prime numbers are essential in various mathematical disciplines, yet their identification continues to challenge researchers. In this brief report, we introduce a unique solution to formulate prime numbers. Using a process of elimination based on a list of the natural numbers as six functions, we systematically eliminate all composite candidates to arrive at the final prime result. In addition, we explore how our model could enhance the efficiency of the Sieve of Eratosthenes in identifying prime numbers.